Decouples client initialisation
