+++ integrations/gitlab-v2/main.py
@@ -7,20 +7,8 @@ from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
 
 from clients.gitlab_client import GitLabClient
 from integration import ProjectResourceConfig
-from utils import ObjectKind
-
-_gitlab_client: Optional[GitLabClient] = None
-
-
-def create_gitlab_client() -> GitLabClient:
-    global _gitlab_client
-    if _gitlab_client is not None:
-        return _gitlab_client
-
-    integration_config: dict[str, Any] = ocean.integration_config
-    base_url = integration_config["gitlab_host"].rstrip("/")
-    _gitlab_client = GitLabClient(base_url, integration_config["gitlab_token"])
-    return _gitlab_client
+from utils.object_kinds import ObjectKind
+from utils.client_factory import create_gitlab_client
 
 
 @ocean.on_start()
