+++ integrations/gitlab-v2/utils/object_kinds.py
